resume_builder/pyproject.toml
[project]
name = "resume-builder"
version = "0.1.0"
description = "Generate .docx Word documents from structured resume data."
requires-python = ">=3.11"
dependencies = [
  "python-docx>=1.1.0",
]

[project.scripts]
resume-builder = "resume_builder.cli:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
